fix: 🐛 runInteractiveQuestions (4.7.6?), refactor: gitmojiTheme
I refactored all of the gitmoji stuff as I thought it was the culprit,
however, looked like an undefined object to `cz` when no `cliAnswers`
were being passed to the runInteractiveQuestions. So broke out
`isDryRun` and `optionsSet` since they are used now in two places. We
can _remove_ `cz` at a later date if this was not right approach to fix.
(I am getting other issues which I think are commitizen related not
git-cz which is breaking previous versions so I am not quite sure what
is going on. Always a good sign, er bad sign, that should have stopped
while I was ahead on all the PRs I wanted to do haha)